WebAcabada-Commit-28/12/2024

+++ database/migrations/2024_12_18_123059_add_image_url_to_products_table.php
@@ -12,17 +12,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('productes', function (Blueprint $table) {
-            $table->string('image_url')->nullable()->after('nom_producte');
+            if (!Schema::hasColumn('productes', 'image_url')) {
+                $table->string('image_url')->nullable()->after('nom_producte');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('productes', function (Blueprint $table) {
-            $table->dropColumn('image_url');
+            if (Schema::hasColumn('productes', 'image_url')) {
+                $table->dropColumn('image_url');
+            }
         });
     }
+
 };
